requirements.txt
Flask==3.1.1
google_api_python_client==2.170.0
gspread==6.2.1
itemadapter==0.11.0
matplotlib==3.10.3
pandas==2.2.3
protobuf==6.31.1
psycopg2_binary==2.9.10
scikit_learn==1.6.1
Scrapy==2.12.0
seaborn==0.13.2
